+++ libs/asset_type_apigw.py
@@ -71,7 +71,7 @@ class APIGW(AssetType):
                 'AuthorizationTypes': self.authorization.types
             }
             if self.public:
-                asset_report['PubliclyAccessible'] = True
+                asset_report['PubliclyAccessible'] = '[red]True[/red]'
             if self.security_issues:
                 self.update_audit_report(asset_report)
         if 'APIGW' not in report[self.location.region]: